Reformatted source using BSD style braces, newline and spacing, except spaces aafter control flow statements. Also resolved many naming conventions issues identified by resharper.

+++ src/AK.F1.Timing/test/Serialization/DecoratedObjectWriterTest.cs
@@ -22,14 +22,14 @@ namespace AK.F1.Timing.Serialization
     public class DecoratedObjectWriterTest
     {
         [Fact]
-        public void can_write_a_null_graph() {
-
+        public void can_write_a_null_graph()
+        {
             Assert.DoesNotThrow(() => Write(null));
         }
 
         [Fact]
-        public void only_graphs_decorated_with_a_type_id_can_be_written() {
-
+        public void only_graphs_decorated_with_a_type_id_can_be_written()
+        {
             Assert.Throws<SerializationException>(() => Write(new object()));
             Assert.Throws<SerializationException>(() => Write('s'));
             Assert.Throws<SerializationException>(() => Write("s"));
@@ -38,7 +38,7 @@ namespace AK.F1.Timing.Serialization
             Assert.Throws<SerializationException>(() => Write((SByte)0));
             Assert.Throws<SerializationException>(() => Write((Int16)0));
             Assert.Throws<SerializationException>(() => Write((UInt16)0));
-            Assert.Throws<SerializationException>(() => Write((Int32)0));
+            Assert.Throws<SerializationException>(() => Write(0));
             Assert.Throws<SerializationException>(() => Write((UInt32)0));
             Assert.Throws<SerializationException>(() => Write((Int64)0));
             Assert.Throws<SerializationException>(() => Write((UInt64)0));
@@ -51,47 +51,48 @@ namespace AK.F1.Timing.Serialization
         }
 
         [Fact]
-        public void graphs_which_contain_circular_references_are_not_supported() {
-
+        public void graphs_which_contain_circular_references_are_not_supported()
+        {
             var graph = new SimpleType();
             // Root to self.
             graph.Value = graph;
             Assert.Throws<SerializationException>(() => Write(graph));
             // Descendent to root.
-            graph.Value = new SimpleType { Value = graph };
+            graph.Value = new SimpleType {Value = graph};
             Assert.Throws<SerializationException>(() => Write(graph));
             // Descendent to self.
             graph.Value = new SimpleType();
             graph.Value.Value = graph.Value;
             Assert.Throws<SerializationException>(() => Write(graph));
             // Descendent to parent.
-            graph.Value = new SimpleType { Value = graph.Value };
+            graph.Value = new SimpleType {Value = graph.Value};
             Assert.Throws<SerializationException>(() => Write(graph));
         }
 
         [Fact]
-        public void the_same_graph_written_on_the_same_stream_is_not_detected_as_a_circular_reference() {
-
+        public void the_same_graph_written_on_the_same_stream_is_not_detected_as_a_circular_reference()
+        {
             var graph = new SimpleType();
 
-            using(var writer = CreateWriter()) {
+            using(var writer = CreateWriter())
+            {
                 writer.Write(graph);
                 writer.Write(graph);
             }
         }
 
         [Fact]
-        public void ctor_throws_if_output_is_null() {
-
+        public void ctor_throws_if_output_is_null()
+        {
             Assert.Throws<ArgumentNullException>(() => new DecoratedObjectWriter(null));
         }
 
         [Fact]
-        public void output_is_not_closed_when_writer_is_disposed() {
-
+        public void output_is_not_closed_when_writer_is_disposed()
+        {
             var output = new MemoryStream();
 
-            using(var writer = new DecoratedObjectWriter(output)) { }
+            using(var writer = new DecoratedObjectWriter(output)) {}
 
             Assert.DoesNotThrow(() => output.Position = 0);
             output.Dispose();
@@ -99,8 +100,8 @@ namespace AK.F1.Timing.Serialization
         }
 
         [Fact]
-        public void write_throws_when_writer_has_been_disposed() {
-
+        public void write_throws_when_writer_has_been_disposed()
+        {
             var writer = CreateWriter();
 
             ((IDisposable)writer).Dispose();
@@ -108,15 +109,16 @@ namespace AK.F1.Timing.Serialization
             Assert.Throws<ObjectDisposedException>(() => writer.Write(null));
         }
 
-        private static void Write(object graph) {
-
-            using(var writer = CreateWriter()) {
+        private static void Write(object graph)
+        {
+            using(var writer = CreateWriter())
+            {
                 writer.Write(graph);
             }
         }
 
-        private static DecoratedObjectWriter CreateWriter() {
-
+        private static DecoratedObjectWriter CreateWriter()
+        {
             return new DecoratedObjectWriter(new MemoryStream());
         }
